+++ frontend/components/editor/types.tsx
@@ -1,0 +1,37 @@
+import { BaseEditor, Descendant } from "slate";
+import { ReactEditor } from "slate-react";
+export type CustomElement = {
+  type: string;
+  align?: string;
+  children: Descendant[];
+  character?: string;
+  url?: string;
+  title?: string;
+};
+export type CustomText = {
+  text: string;
+  bold?: boolean;
+  italic?: boolean;
+  code?: boolean;
+  underline?: boolean;
+};
+
+export type CustomEditor = BaseEditor & ReactEditor;
+
+declare module "slate" {
+  interface CustomTypes {
+    Editor: CustomEditor;
+    Element: CustomElement;
+    Text: CustomText;
+  }
+}
+
+export const HOTKEYS: { [key: string]: string } = {
+  "mod+b": "bold",
+  "mod+i": "italic",
+  "mod+u": "underline",
+  "mod+`": "code",
+};
+
+export const LIST_TYPES = ["numbered-list", "bulleted-list"];
+export const TEXT_ALIGN_TYPES = ["left", "center", "right", "justify"];
