+++ frontend/components/editor/utils.tsx
@@ -1,0 +1,94 @@
+import { Editor, Transforms, Element as SlateElement } from 'slate';
+import { CustomEditor, CustomElement, CustomText, LIST_TYPES, TEXT_ALIGN_TYPES } from './types';
+import isUrl from 'is-url';
+import imageExtensions from 'image-extensions';
+
+export const toggleBlock = (editor: Editor, format: string) => {
+  const isActive = isBlockActive(
+    editor,
+    format,
+    TEXT_ALIGN_TYPES.includes(format) ? 'align' : 'type'
+  );
+  const isList = LIST_TYPES.includes(format);
+
+  Transforms.unwrapNodes(editor, {
+    match: (n) =>
+      !Editor.isEditor(n) &&
+      SlateElement.isElement(n) &&
+      LIST_TYPES.includes(n.type as string) &&
+      !TEXT_ALIGN_TYPES.includes(format),
+    split: true,
+  });
+
+  let newProperties: Partial<CustomElement>;
+  if (TEXT_ALIGN_TYPES.includes(format)) {
+    newProperties = {
+      align: isActive ? undefined : format as CustomElement['align'],
+    };
+  } else {
+    newProperties = {
+      type: isActive ? 'paragraph' : isList ? 'list-item' : format,
+    };
+  }
+  Transforms.setNodes<CustomElement>(editor, newProperties);
+
+  if (!isActive && isList) {
+    const block: CustomElement = { type: format, children: [] };
+    Transforms.wrapNodes(editor, block);
+  }
+};
+
+export const toggleMark = (editor: Editor, format: string) => {
+  const isActive = isMarkActive(editor, format);
+
+  if (isActive) {
+    Editor.removeMark(editor, format);
+  } else {
+    Editor.addMark(editor, format, true);
+  }
+};
+
+export const isBlockActive = (editor: Editor, format: string, blockType = 'type') => {
+  const { selection } = editor;
+  if (!selection) return false;
+
+  const [match] = Array.from(
+    Editor.nodes(editor, {
+      at: Editor.unhangRange(editor, selection),
+      match: (n) =>
+        !Editor.isEditor(n) &&
+        SlateElement.isElement(n) &&
+        (n as CustomElement)[blockType as keyof CustomElement] === format,
+    })
+  );
+
+  return !!match;
+};
+
+export const isMarkActive = (editor: Editor, format: string) => {
+  const marks = Editor.marks(editor) as Partial<CustomText>;
+  return marks ? marks[format as keyof CustomText] === true : false;
+};
+
+export const isImageUrl = (url: string) => {
+  if (!url) return false;
+  if (!isUrl(url)) return false;
+  const ext = new URL(url).pathname.split('.').pop();
+  return imageExtensions.includes(ext!);
+};
+
+export const insertImage = (editor: CustomEditor, url: string) => {
+  const text = { text: '' };
+  const image: CustomElement = { type: 'image', url, children: [text] };
+  Transforms.insertNodes(editor, image);
+};
+
+export const insertMention = (editor: CustomEditor, character: string) => {
+  const mention: CustomElement = {
+    type: 'mention',
+    character,
+    children: [{ text: '' }],
+  };
+  Transforms.insertNodes(editor, mention);
+  Transforms.move(editor);
+};
